fix mahasiswa

+++ system/dosen.php
@@ -213,7 +213,9 @@ class DosenCotroller
         $user = json_decode($this->getById());
         $id = $_POST['id'];
         $file = getImageUpload($user->data->user_id, 'Users.users');
-        $filepath = "../" . $file['data'][0]['path'];
+        if($file['num_rows'] > 0) {
+            $filepath = "../" . $file['data'][0]['path'];
+        }
 
         $sql = "DELETE FROM $this->table WHERE id = ?";
         $params = array($id);
@@ -221,8 +223,8 @@ class DosenCotroller
 
         $delPP = sqlsrv_query($this->conn, "DELETE FROM $this->tableUpload WHERE model_id = ? AND model_name = ?", array($user->data->user_id, 'Users.users'));
         $delUser = sqlsrv_query($this->conn, "DELETE FROM $this->tableUser WHERE id = ?", array($user->data->user_id));
-
-        if (file_exists($filepath)) {
+        
+        if ($file['num_rows'] > 0 && file_exists($filepath)) {
             unlink($filepath);
         }
 
@@ -240,7 +242,7 @@ class DosenCotroller
 
             $file = getImageUpload($_POST['user_id'], 'Users.users');
             $filepath = "../" . $file['data'][0]['path'];
-            if (file_exists($filepath)) {
+           if ($file['num_rows'] > 0 && file_exists($filepath)) {
                 unlink($filepath);
             }
